feat: 삭제시 확인 팝업창 실행

+++ bookstore/src/components/cart/CartItem.tsx
@@ -6,6 +6,7 @@ import { Cart } from '../../models/cart.model';
 import { Title } from '../common/Title';
 import { formatNumber } from '../../utils/format';
 import { useMemo } from 'react';
+import { useAlert } from '../../hooks/useAlert';
 
 interface CartItemProps {
   cart: Cart;
@@ -15,6 +16,7 @@ interface CartItemProps {
 }
 
 function CartItem({ cart, checkedItems, onCheck, onDelete }: CartItemProps) {
+  const { showConfirm } = useAlert();
   //checkedItems 목록에 내가 있는지 판단 = checked
   const isChecked = useMemo(() => {
     return checkedItems.includes(cart.id);
@@ -25,7 +27,9 @@ function CartItem({ cart, checkedItems, onCheck, onDelete }: CartItemProps) {
   };
 
   const handleDelete = () => {
-    onDelete(cart.id);
+    showConfirm('정말 삭제하시겠습니까?', () => {
+      onDelete(cart.id);
+    });
   };
 
   return (
